requirements.txt
pandas
plotly.express
numpy
graphviz
pathlib
tkinter
openpyxl
PyQt5.QtWidgets
sys
